SERVO fixed preprocesser overflow, again
SERVO_US_TO_OCRxA never overflows now.
Also adapted 80-characters-per-line rule for the license.

+++ Servo.c
@@ -9,12 +9,12 @@
  * 
  * Copyright (c) 2018 Sebastian Gössl
  * 
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
+ * Permission is hereby granted, free of charge, to any person obtaining a
+ * copy of this software and associated documentation files (the "Software"),
+ * to deal in the Software without restriction, including without limitation
+ * the rights to use, copy, modify, merge, publish, distribute, sublicense,
+ * and/or sell copies of the Software, and to permit persons to whom the
+ * Software is furnished to do so, subject to the following conditions:
  * 
  * The above copyright notice and this permission notice shall be included in
  * all copies or substantial portions of the Software.
@@ -23,9 +23,9 @@
  * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
  * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
  * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
+ * FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
+ * DEALINGS IN THE SOFTWARE.
  */
 
 
@@ -43,9 +43,9 @@
 #endif
 
 
-#define SERVO_BASE_US (20 * 1000ULL)
-#define SERVO_MIN_US (1 * 1000ULL)
-#define SERVO_MAX_US (2 * 1000ULL)
+#define SERVO_BASE_US (20 * 1000)
+#define SERVO_MIN_US (1 * 1000)
+#define SERVO_MAX_US (2 * 1000)
 #define SERVO_US_PER_SECOND 1000000
 
 
@@ -193,7 +193,7 @@
 
 
 #define SERVO_US_TO_OCRxA(us) \
-    (us * (F_CPU / SERVO_PRESCALER) / SERVO_US_PER_SECOND - 1)
+    ((unsigned long long)us * (F_CPU / SERVO_PRESCALER) / SERVO_US_PER_SECOND - 1)
 
 #define SERVO_BASE_OCRxA SERVO_US_TO_OCRxA(SERVO_BASE_US)
 #define SERVO_MIN_OCRxA SERVO_US_TO_OCRxA(SERVO_MIN_US)
